Move home page into new folder to set prescedent for other pages
